[연경] 맛집 후기 게시판

+++ src/main/java/himedia/hpm_spring/mappers/MountainReviewLikeMapper.java
@@ -4,8 +4,8 @@ import himedia.hpm_spring.repository.vo.MountainReviewLikeVo;
 
 public interface MountainReviewLikeMapper {
 
-//	<select id="selectLikeCount" parameterType="int" resultType="int">	// 특정 리뷰의 좋아요 개수 조회
-    int selectLikeCount(int reviewsId);
+//	<select id="selectLikeCount" parameterType="Long" resultType="int">	// 특정 리뷰의 좋아요 개수 조회
+    int selectLikeCount(Long reviewsId);
 
 //	<select id="isLiked" parameterType="mLikeVo" resultMap="mLikeResultMap">   // 해당 사용자가 해당 리뷰를 좋아요 눌렀는지 여부
     int isLiked(MountainReviewLikeVo mLikeVo);
